// types.ts
export interface Metric {
  title: string;
  value: string;
  change: string;
  changeType: 'increase' | 'decrease';
}